+++ app/SML/models/types/device/supportdevice.cpp
@@ -1,6 +1,6 @@
 #include "supportdevice.h"
 
-SupportDevice::SupportDevice(QString name, QString label, QString index, bool activeState, int mask, QObject *parent) :
+SupportDevice::SupportDevice(QString name, QString label, QString index, bool activeState, byte mask, QObject *parent) :
     Device(name, label, index, activeState, mask, parent)
 {
